requirements/testing.txt
-r base.txt
coverage==4.3.1
selenium==3.0.2
django-allauth==0.30.0
    defusedxml==0.4.1
    oauthlib==2.0.1
    python3-openid==3.0.10
    requests==2.12.4
    requests-oauthlib==0.7.0